Move report button from report.ts to buttons.ts
Moved it onto the same querySelectorPromise as everything else!
