+++ src/buttons.ts
@@ -1,6 +1,8 @@
 import { Program, UserProfileData } from "./types/data";
 import { querySelectorPromise } from "./util/promise-util";
+import { QUEUE_ROOT } from "./types/names";
 import { getCSRF } from "./util/cookie-util";
+import { buildQuery } from "./util/text-util";
 
 //Replace KA's vote button with one that updates after you vote and allows undoing votes
 function replaceVoteButton (buttons: HTMLDivElement, program: Program): void {
@@ -103,6 +105,7 @@ function addLinkButton (buttons: HTMLDivElement, program: Program): void {
 	buttons.insertBefore(document.createTextNode(" "), copyLinkButton.nextSibling);
 }
 
+//Add the number of flags and title text to the program flag button
 function addProgramFlags(buttons: HTMLDivElement, program: Program) {
 	const controls = buttons.querySelector(".discussion-meta-controls")
 
@@ -122,12 +125,32 @@ function addProgramFlags(buttons: HTMLDivElement, program: Program) {
 	}
 }
 
+//Add a button to report the program
+function addProgramReportButton (buttons: HTMLDivElement, program: Program, kaid: string) {
+	if (kaid !== program.kaid) {
+		const reportButton: HTMLAnchorElement = document.createElement("a");
+		reportButton.id = "kae-report-button";
+		reportButton.classList.add("kae-program-button");
+		reportButton.href = `${QUEUE_ROOT}submit?${buildQuery({
+			type: "program",
+			id: program.id.toString(),
+			callback: window.location.href
+		})}`;
+		reportButton.setAttribute("role", "button");
+		reportButton.innerHTML = "<span>Report</span>";
+		buttons.insertBefore(reportButton, buttons.children[1]);
+		buttons.insertBefore(document.createTextNode(" "), reportButton.nextSibling);
+	}
+}
+
 function findOtherButtons (buttons: HTMLDivElement): void {
 	/*Add the kae-program-button class to all other program buttons so we can restyle them */
 	Array.from(buttons.querySelectorAll("a[role=\"button\"]")).forEach(el => el.classList.add("kae-program-button"));
 }
 
 function loadButtonMods (program: Program): void {
+	const kaid:string = (window as any).KA.kaid;
+	
 	querySelectorPromise(".voting-wrap")
 		.then(votingWrap => votingWrap.parentNode)
 		.then(buttons => buttons as HTMLDivElement)
@@ -136,6 +159,7 @@ function loadButtonMods (program: Program): void {
 			addLinkButton(buttons, program);
 			replaceVoteButton(buttons, program);
 			addProgramFlags(buttons, program);
+			addProgramReportButton(buttons, program, kaid);
 		});
 	
 	querySelectorPromise("#child-account-notice")
